feat: move anecdote listing and voting functionality to AnecdoteList component
