+++ part6/redux-anecdotes/src/App.jsx
@@ -1,33 +1,14 @@
-import { useSelector, useDispatch } from "react-redux";
-import { voteAnecdote } from "./reducers/anecdoteReducer";
-import AnecdoteForm from "./components/AnecdoteForm";
+import AnecdoteForm from './components/AnecdoteForm'
+import AnecdoteList from './components/AnecdoteList'
 
 const App = () => {
-  const anecdotes = useSelector((state) => state); //accessing the state from the store
-  const sortedAnecdotes = [...anecdotes].sort((a, b) => b.votes - a.votes);
-  const dispatch = useDispatch(); //dispatching actions to the store
-
-  //handling the voting of an anecdote
-  const vote = (id) => {
-    dispatch(voteAnecdote(id));
-  };
-
   return (
     <div>
       <h2>Anecdotes</h2>
-      {sortedAnecdotes.map((anecdote) => (
-        <div key={anecdote.id}>
-          <div>{anecdote.content}</div>
-          <div>
-            has {anecdote.votes}
-            <button onClick={() => vote(anecdote.id)}>vote</button>
-          </div>
-        </div>
-      ))}
-      <h2>create new</h2>
+      <AnecdoteList />
       <AnecdoteForm />
     </div>
-  );
-};
+  )
+}
 
-export default App;
+export default App
